Find Existing Instances of BundledSQLiteConnection
The inspector currently doesn't register BundledSQLiteConnection's that are already open when the inspector attaches.

We need to find existing instances in memory similar wo what we do with SQLiteDatabase.

This means we can't relly on SQLiteConnectionWrapper to handle close
and prepare hooks because we can't replace an already existing  connection with a wrapper. Instead, we drop the wrapper and hook close and prepare with ART Tooling.

Testing this specific scenario requires refactoring of the test infrastructure. Will do that in a subsequent cl.

Fixes: 434123777
Test: Manually and existing
Change-Id: I63e5c52716528a2811cb1bf2700b0f8f964c2056