Re-enable -force-snapshot-load for emulators on Linux
The emulator can delete its snapshot upon a failed boot, even when
launched with the `-read-only` flag. This behavior can cause test
instability, especially when multiple AVD instances rely on the same
snapshot.

This change restores the `-force-snapshot-load` flag for emulators on
Linux, which prevents the snapshot from being deleted on a failed boot.

The flag was previously disabled globally (I18e410bf7f) because the
older emulator version used on macOS build bots does not support it. It
remains disabled for macOS until the emulator is updated.

Bug: n/a
Test: MultipleEmulatorTest
Change-Id: Ia2b273b4454c8efb3803f91d8f54eee55d687e2c