adblib: fix resumeProcess for app_info
When app_info is supported, the implementation of
`JdwpProcess.resumeProcess` can be straightforward
and directly open a JDWP session to the process, i.e.
there is no need to dispatch the command to all
running instances of `adblib` via the Inventory Server,
because there should be no single instance using
a JDWP connection to the process.

An end-2-end test is added, a large CL will come later
to add an end-2-end test for devices that don't support
`app-info` (it is more complex, as we need to update
FakeAdbServer to better emulator how the `WAIT` packet
is handled).

Bug: 333389175
Test: Included
Change-Id: I40ee760597b4db82ef1b0e9ceaded288bf2c0d70