Fix failing JourneysConnectedTest for gradle 9.0 by removing dependency on a gradle specific string message.
Gralde 9.0 changes the behaviour for when no tests are executed, see
https://docs.gradle.org/8.7/userguide/upgrading_version_8.html#test_task_fail_on_no_test_executed.
The new error message says:
"""
Execution failed for task ':app:validateDebugJourneysTest'.
> java.util.concurrent.ExecutionException:
org.gradle.api.tasks.testing.TestExecutionException: There are test
sources present and no filters are applied, but the test task did not
discover any tests to execute. This is likely due to a misconfiguration.
Please check your test configuration. If this is not a misconfiguration,
this error can be disabled by setting the 'failOnNoDiscoveredTests'
property to false.
"""
instead of "No test executed"

Hence, it's more stable to not rely on a specific gradle message, rather
we can just make sure that the specific test name does not occur in the
output.

Test: JourneysConnectedTest
Bug: 437851001
Change-Id: I1bd291d4c8b24c8b6c1fceae0f6fca97143b44b8