Fix managed device task resolution in KMP projects
When using GMD in a KMP Android library project, the build would fail
with a TaskDependencyResolveException.

This occurred because the `AvdComponentsBuildService`, a required
dependency for GMD to function, was not being registered by the KMP
Android plugin.

This patch registers the service, resolving the exception and allowing
managed device tests to run successfully in KMP projects.

Test: KmpManagedDeviceTest
Fixes: 436887358
Change-Id: Idc2b61a498e401944fe297107fe73cb67320c1e0