Do not include R.jar to the Test's classes
When a project test options include android resources for host
tests, the R.jar must be made available for compiling and should
also be available in the tests' classpath. However, the R.jar
should not be passed to the Test class inputs for tests classes.
Instead, it must only be made available to the test runtime classpath.
The AndroidUnitTest task was setting up the runtime classpath correctly
but the R.jar was also added to the inputs for test classes.

Test: Added a regression test
Bug: 411739086
Change-Id: I8a90476248aa20a987d76a5af6ba60f6633a8398