Fork AmInstrumentationParser from UTP core into AGP
As part of the effort to remove AGP's dependency on the Unified Test
Platform (UTP), this change forks the `AmInstrumentationParser`.

This parser is essential for processing the output of `am instrument`
commands. Since the original implementation has few dependencies on
other UTP components, it can be copied directly into the AGP codebase to
eliminate the dependency.

This change introduces the parser and its required helper classes,
`AmInstrumentationListener` and `TestTimeTracker`, along with a
comprehensive test suite.

This is an interim solution. The long-term plan is to replace this
forked implementation with a new one based on adblib.

Test: AmInstrumentationParserTest
Bug: 434015775
Change-Id: Iced0cb36c820d2a526af87d16c0619da1f223ec7