Re-enable the old `ThreadDetector` by default
The new thread detector has lingering problems with divergence that
cause some builds to timeout. This change separates the new detector's
`issue` from the old intra-procedural one and re-enables the old one by
default. The new detector still replaces the old
`WrongThreadInterprocedural`, because that issue is disabled by default,
and the new one still should improve on it in practice 🤞.

Bug: 440375253
Test: existing
Change-Id: I7ec425a53dc3d3001ee597db41c3e67845cfa283