Expose image diff percentage from ImageVerifier
Refactored `ImageVerifier` to return a `VerificationResult` data class
from the `verify` method. This class encapsulates the `ImageDiffer.DiffResult`
and the calculated `diffPercent` string. This change removes the stateful
`lastDiffPercent` property, addressing design concerns.

`PreviewScreenshotDescriptor` now uses the `VerificationResult` to
access the diff percentage and reports `PreviewScreenshot.diffPercent`
unconditionally in the `finally` block. An `ImageComparisonAssertionError`
is thrown only if the result type is `Different`.

Bug: 442358523
Test: Added unit tests to ImageVerifierTest.kt
Change-Id: I3b751f557f1a84497f53b313f82ada4b117d0778