Update library desugaring databases
When lint runs as part of AGP, AGP passes R8's desugaring configuration
files to it such that it knows exactly which APIs are desugared and
which are not. But lint also has a fallback database it uses when it's
not running as part of AGP, e.g. from unit tests, or simple command line
invocations (such as in g3).

This CL updates these databases to the latest versions of these --
com.android.tools:desugar_jdk_libs_configuration_nio:2.1.5 and
java -cp $ANDROID_HOME/cmdline-tools/latest/lib/r8.jar com.android.tools.r8.BackportedMethodList

Bug: N/A
Test: Existing
Change-Id: If4468bb76bd1455925f33fead9c99a02488d543b