[fused lib] Enable fused library support by default
Sets Fused Library support as enabled by default and sets the feature
stage to Enforced as there are no longer production users of the option,
Disabling has no effect, as the same can be achieved by not applying the
plugin. The only real impact is not showing the experimental plugin
warning to users.

Bug: 346984830
Test: Existing
Change-Id: If6338099d3c57271689e41d44e8b52912e34a334