Fix version lookup
Looking up gradle artifacts in the org.jetbrains.kotlin namespace was
not working because there's a match in the maven.google.com repository
(so it doesn't check maven central). This should not be a match since
the org.jetbrains.kotlin namespace on maven.google.com was only
temporarily used for KSP related libraries, which have long ago moved.

Bug: Not filed
Test: Unit tests updated
Change-Id: Ia522cf78e4b482c00840ca21c79560c3a7ff2fd5