New RichVersion factory.
The semantics of version specifiers for dependencies in POM files are
not the same as the semantics of version specifiers for Gradle.
Define a RichVersion#fromPomVersion factory to convert the common
cases correctly, and less-common cases approximately correctly.

The definition of the semantics of version specifiers for dependencies
in POM files is is actually a list of examples rather than anything
clearly specified, so this implementation comes with some caveats; it
is implemented with reference to the documentation, such as it is, and
also from a careful read of the implementation in the Maven project
itself.

This factory preserves the Gradle semantics of an interval with an
open upper endpoint, in order to minimize the number of entities this
factory will generate with no printed identifier representation.
Otherwise, for valid Maven dependency verison specifiers, it should
produce Gradle RichVersions with identical semantics; for invalid
Maven version specifiers, the results might be minimally helpful.

Use this new factory in GoogleMavenRepository and
GoogleMavenRepositoryV2.

Bug: 446738978
Test: added, adjusted
Change-Id: I6a75a1ebb9e12c76b2141b5cdd1660962ef9d17a