Add support for purposeString validation in PPD lint rule.
Adding support to a new free-form purpose type. When this is required,
developers would need to define a purposeString attribute on
uses-permission, defining a string resource. The lint rule has been
extended such that if the purposeString attribute is missing or not of
the right type, an error is reported.

If a permission requires specific purpose and free-form purpose and
don't have either, the reporting still happens only once but the error
messages of both types are combined in the message.

The validation of the purposeString text will happen in a different CL.
In fact, we will register a new ID so in case of performance issues, it
can be suppressed without impacting error detection of other purpose
issues.

Additional changes:
1. To avoid verbosity when combining multiple messages, the error
   messages have been shrinked. Added more detail to the MISSING_PURPOSE
   issue explanation.
2. Updated all tests by deleting the grade configuration and encoding
   the version info in the manifest due to a lint bug.

Bug: 442571488
Test: PurposeDeclarationDetectorTest
Change-Id: Id94569cabe3c1580ac4f2a9b25c22624d1d17eef