Fixes sources.kotlin.addGeneratedSourceDirectory() when using kotlin-android
When using kotlin-android, tasks generating sources to kotlin, are not executed and created because kotlin-android plugin does not consume them. In other words the API does not have any effect. If instead sources.java.addGeneratedSourceDirectory() is used, generation happens.
This change updates getSourceFolders() API - which is what kotlin-android uses to get sources from AGP - to return kotlin sources as well.

Note: Things work as expected with built in kotlin already

Bug: 446220448
Test: GeneratedKotlinSourcesAccessViaOldVariant.kt
Change-Id: I120713c988f67f84a451cf62c0b0f60e50ab811d