Set HOME to TEST_TMPDIR for tests
The --sandbox_fake_username option, which is enabled in the common
configuration, changes the user to "nobody" for tests running in the
sandbox.

This causes tests that rely on a valid HOME directory (such as those
using adb) to fail.

This commit explicitly sets HOME="$TEST_TMPDIR" using the --test_env
flag. This ensures that all sandboxed tests have a HOME directory
defined, allowing them to run correctly.

presubmit-test: default
Test: existing
Fixes: 454688286
Change-Id: I62ce5bae783c6c43f4a2aed0b3d03246fa338b01