Locally cache analysis results of some sub-expressions
The way the analysis was handling operators that support `multiResolve`
was prone to exponential blowup: for each method that the operator
resolved to, we re-analyzed the same arguments all over again. A small
program with nested uses of operators (e.g. a += { b += ... }), where
`+=` resolved to 10 different methods, took 6+ minutes to analyze.

A standard refactoring would be just calling `loop` on each
sub-expression once, and passing the sub-results (instead of the
sub-expressions themselves) to `callMethod`. But UAST might resolve the
same operator to different methods (e.g. `+=` to either `plusAssign` or
`plus`), requiring different sub-expressions and sub-sub-expressions,
and calling `loop` on each of them ahead of time would still cause some
duplicate work. We therefore maintain a local cache, and recurse through
it for calls that might be duplicated elsewhere.

Bug: n/a
Test: existing for behavior, manually for performance
Change-Id: Ie0fd1a9b43b819e9395dcebca3bc2dffc0837c4a