Introduce LocalTestProjectSpec to configure on-disk projects
When multiple test classes use the same on-disk test projects,
it's convenient to have a reusable logic that configures the build
files for the test project's sub-projects. This commit
introduces a mechanism to do just so. The spec contains
both a project name (e.g. "basic"), and a config action
for GradleBuildDefinition to write all the build files.

This commit also converts all the project that used
"basic" to this new mechanism. In order to do this,
we had to change how the basic project is architectured
as the fixture does not support root level android
projects. The on-disk version of build.gradle has
been removed since all projects using it where converted.

Finally, this commit also brings in needed fixes to the
GradleRule fixture in order to convert the basic tests
- Support for signingConfigs block
- Supoprt for assigning a signingConfig instance to
  a property (using for BuildType and Flavor)
- Support for publishing.singleVariant with a lambda
  parameter (or any other method where the lambda is
  not the first and only param)
- Support for declaring flatDir repositories in
  sub-project

Bug: N/A
Test: existing tests
Change-Id: Ic8d6d302978e48827ef37073ab1ad00ce301a11b