Do not use the highest installed build target if not specified in the projects
Prior to this change, if a project did not specify a `compileSdkVersion`,
lint would default to using the highest installed SDK platform.
This could lead to issues where lint would analyze the project with a
higher `compileSdkVersion` than what was actually being used,
potentially leading to incorrect warnings.

This change refactors the logic for looking up the compile target to
differentiate between explicitly specified targets and defaulted targets.
The lint CLI now uses this information to pick a build target, prioritizing
explicitly specified targets and only falling back to the highest installed
platform if no project specifies a `compileSdkVersion`.

This ensures that lint's analysis is more accurate and reflects the actual
build environment of the project.

Bug: 444447002
Test: tests added
Change-Id: Idbe6f50bb6ef54d2f110aab25cf3dcd487fef59f