add safeguard against using Provider<> for proguard files DSL APis
it was decided to not support setting a Provider<> to android
source set as it is not possible for the IDE to know if the file
or directory should be read-only or not. Instead users should use
the variant API.

provided a message guiding the user to the proper APIs.

Test: Added a regression test.
Fixes: N/A
Relnote: Do not use Provider<> for proguard files APIs in the DSL, instead use the Variant API equivalent APIs.
Change-Id: Ibb184d62a6963a4240e1dfd74bed3dbf4e9100f2