[pagealign] Add RELRO checks to ApkViewer
Checks whether RELRO sections of the ELF header are misaligned and
likely to cause crashes on 16 KB devices.

Warning message in APK viewer:
    go/ss/BmrJstxB8pEqq6j
This check won't gate Play Store upload so we don't give the Play Store
warning if there are only RELRO issues.

There might be multiple issues. We show them all in the summary pane:
    go/ss/3ua8LDrVJYUKj3S

The logic for the checks are aligned with the logic in the 'aac' tool
does.

    cs/h/googleplex-android/platform/superproject/main/+/main:system/memory/libmeminfo/libelfutils/tools/aac/aac.cpp

Bug: 455976781
Test: Unittested. Validated against real world.
Change-Id: Iafb40870fe8a38748a77dbe3372b1b9c83d3d293