Remove useless TemporaryProjectModification in MessageRewriteTest
Change I4937eae497a841786cd7d2c0e10bc597020fd644, in converting from
Groovy, also changed a ClassRule to a regular Rule, meaning that each
test runs with its own instance of the project rule.  This further
means that changes to the project are not visible to any other test.

Bug: none filed
Test: existing
Change-Id: I1341f68ad8357c08c5dbd58ec0c27444767c31ea