Reduce newDsl=false in native AGP integration tests
The only reason for having USE_NEW_DSL be false in these two test
classes is small fragments of assertion directly in the build file.

For CmakeBasicProjectTest, rewrite the assertion in a newer API, and
delete the meta-test which apparently does nothing but test that the
test-generated build file contained the right string.

For NdkBuildBuildSettingsTest, remove the build file assertion, and
instead run the task directly in a new test.

Fixes: 463934800
Test: these are tests
Change-Id: Ie4f9a61f6307eb58507889aa3e765031c0ce04d3