Adapt VariantFilteringTest for newDsl
Move tests which don't require the newDsl flag to be `false` to
AndroidComponentsFilteringTest, which seems to be essentially testing
the same kind of thing.  Trivially translate the given / android /
androidComponents / expect mini-Dsl, but all the tests should be
semantically identical.

The remaining two tests in VariantFilteringTest both need USE_NEW_DSL
set to `false`, so do that.

Bug: none filed
Test: existing
Change-Id: I48c55ce52656bc211010024275c2d67f214c5c36