Special cases for tests of the interaction between old and new API
Those tests by their very nature require running with the USE_NEW_DSL
flag set to false.  Separate out tests of interaction misuse from the
others in VariantApiMisuseTest, adding an entry to the allowlist in
order to preserve the existing tests wholesale (with only the addition
of the flag).

Likewise, add the flag to OldAndNewVariantAPIInteractionTest.

Bug: none filed
Test: none added, none removed
Change-Id: Ibc2441ee4533cc4f5bd391cdd4695c667d61f580