[res] Explictly mark relative and absolute path encoding
This change adds ResourcePathEncoding that is now required for merge
and compile resource requests so it is easier to identify where absolute
paths are still being encoded.

This is a precursor to I584fde15fa2b8a79e209cfb62ff440eaff4644f0 where
absolute paths are no longer writen to compiled library resources.

Bug: 278255388
Bug: 398173037
Test: Existing
Change-Id: I08a77cf99a39f78e55155135c7923f13a177294a