use separate book keeping for user added classes
users are using the public CLASSES scoped artifacts to add new
directories of class files. Such directories should be used to
create the project classes.jar. However, such classes.jar are not
created from the CLASSES scoped artifacts as it may contain the
instrumented classes but usually assembled manually by combining
KOTLINC and/or JAVAC outputs, plus other ones.

The user create class files should be part of these outputs to create
the final module classes.jar which forces us to use a separate book
keeping.

One relatively easy way to do that (outside of creating yet another
public artifact type to register CLASSES producers) is to intercept
the user's invoked toAppendTo methods calls and store the provider
separately. This internal storage called is a dedicated field on
the scoped artifacts container. This could be improved in the
future to avoid special casing the container.

One of the difficulty is that even though CLASSES are tagged as
directory, none of the scoped API are actually checking that users
are passing directories. In fact we have a recipe that is passing
both a folder and a jar so this special handling with a special
field was required to handle the absence of type when appending
CLASSES.

the module classes.jar and will also be republished as part of the
CLASSES artifact for the PROJECT scope.

Test: Added a integration test, ensuring the pre and post compilation
hooks are part of the module classes.jar published artifact.
Bug: 444641424

Change-Id: If2cf7d6b1ea41fbd8ee1d377299a0244a77f9bf1