abstract configuration for kotlin related tasks.
The current CreationAction for kotlin related tasks are relying
on a CreationConfig instance which assumes it is dealing with a
Component. That's not the case for TestSuite. Therefore, abstract
the required information from the ComponentCreationConfig into
a new interface BuiltInKotlinCreationConfig which instances will
for now just delegate back to the ComponentCreationConfig.

A subsequent CL will introduce new implementations of this
interface suitable for test suites compilation.

No functional changes.
Test: Existing
Bug: N/A

Change-Id: Ide6c0d82fd48840688d096a7c8a6873611625ddc