Ensure GradleInstallationManager.getGradleJvmPath is invoked from Job
With IntelliJ 2025.1, the SdkLookupProvider.resolveGradleJvmInfo
implementation was migrating to coroutines making implementation of
GradleInstallationManager.getGradleJvmPath (non-coroutines) to invoke
runBlockingCancellable when doing a lookup for GradleJvmInfo which
was introduced by commit 63703d06a954.

The change SdkLookupProvider.resolveGradleJvmInfo makes completely
sense, since resolving the project Gradle JDK configuration might
require to read files to obtain the specified JDK path. Calling from
EDT will now lead to the below log error which is considered an
exception during test execution, but is also reported to crash logs:
    "There is no ProgressIndicator or Job in this thread,
    the current job is not cancellable."

However, to enforce proper behaviour a new AS extension was defined
to override GradleInstallationManager and deprecate the
'getGradleJvmPath()' in favor of suspend 'resolveGradleJvmPath()'
function ensuring that a job is always present.

Fixes: 452882350
Test: Existing
Change-Id: If13f2b7e3ca694792fcfa4a59f970d143f33f2a1