Remove problematic iml module
This code is imported from google3 and relies on being built
primarily with bazel, since it uses autovalue. This means we
can't easily also have an iml module for it, since (as seen here)
you end up having to have the module depend on a library containing
its own code, resulting eventually in duplicate classes.

Test: DuplicateClassesTest
Bug: 476427677
Change-Id: Id2fcfaed154505a956430dbc94d2596b5a010119